separate plots with/without dropsondes
